Create functionality to join users in chat rooms
Created more functions in room-controller.js; (enterRoom, exitRoom,
addOnlineUser and removeOnlineUser)
	- enterRoom: to check if the user is not already in this room.
		(1 promise) Calls addOnlineUser().
	- exitRoom : to check if the user is already in the room.
		(1 promise) Calls removeOnlineUser().
	- addOnlineUser: adds the user to room.onlineUsers,
		and saves the room (1 promise).
	- removeOnlineUser: Removes the user from room.onlineUsers,
		and saves the room(1 promise).
When the user first enters the room, a check is made inside enterRoom(),
to see if the user is not already in this room.
Then, addOnlineUser is called.
When the user exits, exitRoom() is called. Then the check to see if the
user is present in the room, is in removeOnlineUser(), which is called
immediately. The check is in the latter function to avoid any
unprediected errors.

Updated room-model.js to contain a list of connected users
	(user.id, and user.name).

Routers are updated properly to use the newly created functionality
from room-controller.js

views/user/profile.ejs is updated, to show the correct credentials
of the user.
Editing user information is not fully functional yet
(ajax call to be written).