Divide the main namespace in socket-events to multiple namespaces
Grabs a list of all created rooms from the database, converts
them to namespaces, and assigns different events for these namespaces

Exports sharedSession and sessionSetup to socket-events,
for them to be applied on all of the namespaces.

This approach is flawed. Because newly created rooms wouldn't have
any events assigned to them.
It must be changed -instead- to the main namespace. And users
can be connected to channels.