Rewrite Room Controller and convert its methods into promises
Converts most of the functionalities in Room Controller into
promises.

Database calls are reduced by keeping a room and a user
objects in memory in(socket-event.js).

Socket-events redesigned. When a user connects to the main
namespace, he is directed to a separate channel identified
by the name of the the user's currently visited room.

A beep sound plays when a user gets a message, without
having the browser window focused at.